requirements.txt
optimum-intel
openvino
diffusers
onnx
fastapi
uvicorn[standard]
python-multipart
Pillow
